Department user complated

+++ app/routers/sub_task.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, WebSocket, HTTPException
+from fastapi import APIRouter, Depends, HTTPException
 from sqlalchemy.orm import Session
 from ..schemas import SubtaskCreate, SubtaskRead, SubtaskUpdate
 from app.crud import create_subtask, update_subtask, read_subtasks, delete_subtask
@@ -9,26 +9,40 @@ from ..auth import get_current_user
 from typing import List
 
 def check_subtask_exists(db: Session, subtask_id: int):
-    if db.query(Subtask).filter(Subtask.id == subtask_id).first() is not None:
-        return HTTPException(status_code=404, detail="Subtask not found")
+    return db.query(Subtask).filter(Subtask.id == subtask_id).first() is not None
+        
     
 
 router = APIRouter(prefix="/subtasks", tags=["subtasks"])
 
-@router.post("/", response_model=SubtaskRead, dependencies=[Depends(require_role(RoleEnum.company_admin, RoleEnum.department_manager))])
+@router.post("/", response_model=SubtaskRead)
 def create(subtask_in: SubtaskCreate, db: Session = Depends(get_db)):
+    user = get_current_user()
+    if user.id != subtask_in.task.department.manager_id or user.role != RoleEnum.company_admin:
+        raise HTTPException(status_code=403, detail="You can only create subtasks for your own tasks")
     return create_subtask(db, subtask_in)
 
 @router.get("/{task_id}", response_model=List[SubtaskRead])
 def list_all(task_id: int, db: Session = Depends(get_db)):
+    if db.query(Subtask).filter(Subtask.task_id == task_id).count() == 0:
+        raise HTTPException(status_code=404, detail="No subtasks found for this task")
     return read_subtasks(db, task_id)
     
-@router.put("/{subtask_id}", response_model=SubtaskUpdate, dependencies=[Depends(require_role(RoleEnum.company_admin, RoleEnum.department_manager))])
+@router.put("/{subtask_id}", response_model=SubtaskUpdate)
 def update(subtask_id: int, subtask_in: SubtaskUpdate, db: Session = Depends(get_db)):
-    check_subtask_exists(db, subtask_id)
+    user = get_current_user()
+    if user.id != subtask_in.task.department.manager_id or user.role != RoleEnum.company_admin:
+        raise HTTPException(status_code=403, detail="You can only update subtasks for your own tasks")
+    if not check_subtask_exists(db, subtask_id):
+        raise HTTPException(status_code=404, detail="Subtask not found")
     return update_subtask(db, subtask_id, subtask_in)
 
-@router.delete("/{subtask_id}", dependencies=[Depends(require_role(RoleEnum.company_admin, RoleEnum.department_manager))])
+@router.delete("/{subtask_id}")
 def delete(subtask_id: int, db: Session = Depends(get_db)):
-    check_subtask_exists(db, subtask_id)
+    user = get_current_user()
+    subtask_in = db.query(Subtask).get(Subtask.id == subtask_id, None)
+    if user.id != subtask_in.task.department.manager_id or user.role != RoleEnum.company_admin:
+        raise HTTPException(status_code=403, detail="You can only delete subtasks for your own tasks")
+    if not check_subtask_exists(db, subtask_id):
+        raise HTTPException(status_code=404, detail="Subtask not found")
     return delete_subtask(db, subtask_id)
